+++ test/EXAMPLE -- Visual Experiment.py
@@ -13,11 +13,11 @@ def trial(word, pos):
 
     key, rt = waitForResponse({key_.F: 'Left', key_.J: 'Right'}) # Waiting for pressing 'F' or 'J'
     if key == word:
-        alertAndGo('Correct!', 1000)  # Display something in 1s
+        alertAndGo('Correct!', 1)  # Display something in 1s
     else:
-        alertAndGo('Wrong!', 1000)
+        alertAndGo('Wrong!', 1)
 
-    show(500)  # Pause (Keep displaying in 500ms)
+    show(0.5)  # Pause (Keep displaying in 0.5s)
 
 
 instruction(shared.setting['instruction1'])