+++ test/rsvp.py
@@ -10,4 +10,4 @@ start()  # Initiate the experiment environment
 
 for w in 'ABCDE12345':
     drawText(w)  # Draw something on the canvas(not the screen)
-    show(200)  # Display current canvas on the screen, and keep for 200ms
+    show(0.2)  # Display current canvas on the screen, and keep for 0.2s