+++ test/scaffold.py
@@ -8,7 +8,7 @@ sys.path = ['../../']+sys.path
 from expy import *  # Import the needed functions
 start()  # Initiate the experiment environment
 
-Get user input until "ENTER" pressed, then give it to a variable
+# Get user input until "ENTER" pressed, then give it to a variable
 something = getInput('enter something:')
 
 # Show something until "RETURN" pressed
@@ -18,10 +18,10 @@ alert('You just entered "%s".\nPlease press RETURN to continue.' %something)
 instruction(['This is the first page of instruction>', 'second page>', 'last page\nPress SPACE to quit the instruction'])
 
 # Suspend the experiment and ask participant to rest, until "SPACE" pressed
-restTime()  
+restTime()
 
 # Show something during a limited period, and continue
-alertAndGo('Show something for 3000ms', 6000)
+alertAndGo('Show something for 3s', 3)
 
 # Show something during a limited period, and quit the program
-alertAndQuit('Show something for 3000ms, and quit')
+alertAndQuit('Show something for 3s, and quit')