+++ test/show_picture.py
@@ -11,21 +11,21 @@ start()  # Initiate the experiment environment
 '''General usage'''
 # Draw a picture on the canvas center
 drawPic('data/demo.jpg')
-show(3000)  # Display current canvas
+show(3)  # Display current canvas
 ''''''
 
 # Draw a zoomed picture on the canvas center
 drawPic('data/demo.jpg', w=400, h=300)
-show(3000)  # Display current canvas
+show(3)  # Display current canvas
 
 # # Draw a zoomed picture on the canvas center
 # drawPic('data/demo.jpg', w=300, h=400, rotate=90)
-# show(3000)  # Display current canvas
+# show(3)  # Display current canvas
 
 # Draw a zoomed picture on the canvas, and move it
 drawPic('data/demo.jpg', w=400, h=300, x=0.5, y=0.5)
-show(3000)  # Display current canvas
+show(3)  # Display current canvas
 
 # Draw a zoomed picture on the canvas, and move it
 drawPic('data/demo.jpg', w=400, h=300, x=0.5, y=0.5, anchor_x='left',anchor_y='center')
-show(3000)  # Display current canvas
+show(3)  # Display current canvas